Continue on errors for some CI steps
We want the subsequent job step to run even if the unit tests fail,
but we want the overall job to be marked as a failure.
So we use conditionals on the subsequent steps,
rather than "continue-on-error" settings.

Installing MOPAC, however, is allowed to fail entirely
because it will always do so on pull requests from forks
(that don't have access to the secrets).
And it times out after a minute, as it has a habit
of hanging if you get it wrong.