Remove util from arkane documentation, as it doesn't exist any more.
This was moved to an ess factory method in be9c22ed723676fd954b06a429a122c3a47f3ea4
but I'm not sure how to add in the replacement into this index.
For now I'm trying to just make the documentation not crash or give missing 
references. Someone will have to make it "complete" later.