set default ascend=False
In general while ascend=True is the most proper way to do this in tests it gives pretty much the same errors as ascend=False, but it takes drastically longer
to compute because it has to refit the BMs when it ascends. In most cases ascend=False should be sufficient.